[#78] frequency-calculator: fix + emulib jar name

+++ src/main/java/net/emustudio/emulib/plugins/cpu/FrequencyCalculator.java
@@ -43,7 +43,7 @@ public class FrequencyCalculator implements CPUContext.PassedCyclesListener, Clo
             synchronized (this) {
                 long endTime = System.nanoTime();
                 long time = endTime - startTime;
-                long executedCycles = cycles.get();
+                double executedCycles = cycles.get();
 
                 if (executedCycles > 0) {
                     frequencyKhz = (float) (executedCycles / (time / 1000000.0)); // per millisecond = kHz
